GV2-123 | Updated Icons: * Updated SocialIcon.tsx * Updated Icons.stories.tsx * Moved icons to IconList folder
