+++ src/components/Chip/Chip.stories.tsx
@@ -3,7 +3,7 @@ import {ComponentStory, ComponentMeta} from '@storybook/react-native';
 import Chip from '.';
 import CenterView from '../../storybook/preview/CenterView';
 import {Platform} from 'react-native';
-import {ArrowDownSmallIconDeprecated, CheckIconDeprecated} from '../Icons';
+import {CheckIconDeprecated} from '../Icons';
 import pkg from './package.json';
 
 export default {
@@ -62,7 +62,7 @@ DisabledOutlined.args = {
 export const WithIcons = Template.bind({});
 WithIcons.args = {
   children: 'Icons',
-  iconRight: <ArrowDownSmallIconDeprecated style={{marginHorizontal: 3}} />,
+  iconRight: <CheckIconDeprecated style={{marginHorizontal: 3}} />,
   iconLeft: <CheckIconDeprecated style={{marginHorizontal: 2}} />,
 };
 
@@ -75,7 +75,7 @@ LeftIcon.args = {
 export const RightIcon = Template.bind({});
 RightIcon.args = {
   children: 'RightIcon',
-  iconRight: <ArrowDownSmallIconDeprecated style={{marginHorizontal: 3}} />,
+  iconRight: <CheckIconDeprecated style={{marginHorizontal: 3}} />,
 };
 
 export const OnlyIcon = Template.bind({});
