+++ src/components/Icons/IconList/CountIcon.tsx
@@ -1,4 +1,4 @@
-import {createCommonIcon} from './utils';
+import {createCommonIcon} from '../utils';
 
 const iconVariants = {
   minus: `<svg width="24" height="24" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
