+++ src/components/Icons/index.tsx
@@ -1,32 +1,30 @@
-export {default as InfoIcon} from './InfoIcon';
-export {default as CheckIconDeprecated} from './CheckIconDeprecated';
-export {default as CheckIcon} from './CheckIcon';
-export {default as ArrowDownIconDeprecated} from './ArrowDownIconDeprecated';
-export {default as ArrowRightIconDeprecated} from './ArrowRightIconDeprecated';
-export {default as EyeIconDeprecated} from './EyeIconDeprecated';
-export {default as EyeSlashIconDeprecated} from './EyeSlashIconDeprecated';
-export {default as ArrowDownSmallIconDeprecated} from './ArrowDownSmallIconDeprecated';
-export {default as CloseIconDeprecated} from './CloseIconDeprecated';
-export {default as ArrowIcon} from './ArrowIcon';
-export {default as PointArrowIcon} from './PointArrowIcon';
-export {default as QuestionIcon} from './QuestionIcon';
-export {default as EyeIcon} from './EyeIcon';
-export {default as CalendarIcon} from './CalendarIcon';
-export {default as CloseCircleIcon} from './CloseCircleIcon';
-export {default as CreditCardIcon} from './CreditCardIcon';
-export {default as MoreIcon} from './MoreIcon';
-export {default as MicIcon} from './MicIcon';
-export {default as ProfileCircleIcon} from './ProfileCircleIcon';
-export {default as CountIcon} from './CountIcon';
-export {default as LoadIcon} from './LoadIcon';
-export {default as EditIcon} from './EditIcon';
-export {default as HomeIcon} from './HomeIcon';
-export {default as FilesIcon} from './FilesIcon';
-export {default as ProfileIcon} from './ProfileIcon';
-export {default as DocumentIcon} from './DocumentIcon';
-export {default as DeleteIcon} from './DeleteIcon';
-export {default as ContentCopyIcon} from './ContentCopyIcon';
-export {default as PaymentMethodIcon} from './PaymentMethodIcon';
-export {default as SocialIcon} from './SocialIcon';
-export {default as CloseIcon} from './CloseIcon';
-export {default as SearchIcon} from './SearchIcon';
+export {default as InfoIcon} from './IconList/InfoIcon';
+export {default as CheckIconDeprecated} from './IconList/CheckIconDeprecated';
+export {default as CheckIcon} from './IconList/CheckIcon';
+export {default as ArrowDownIconDeprecated} from './IconList/ArrowDownIconDeprecated';
+export {default as EyeIconDeprecated} from './IconList/EyeIconDeprecated';
+export {default as EyeSlashIconDeprecated} from './IconList/EyeSlashIconDeprecated';
+export {default as CloseIconDeprecated} from './IconList/CloseIconDeprecated';
+export {default as ArrowIcon} from './IconList/ArrowIcon';
+export {default as PointArrowIcon} from './IconList/PointArrowIcon';
+export {default as QuestionIcon} from './IconList/QuestionIcon';
+export {default as EyeIcon} from './IconList/EyeIcon';
+export {default as CalendarIcon} from './IconList/CalendarIcon';
+export {default as CloseCircleIcon} from './IconList/CloseCircleIcon';
+export {default as CreditCardIcon} from './IconList/CreditCardIcon';
+export {default as MoreIcon} from './IconList/MoreIcon';
+export {default as MicIcon} from './IconList/MicIcon';
+export {default as ProfileCircleIcon} from './IconList/ProfileCircleIcon';
+export {default as CountIcon} from './IconList/CountIcon';
+export {default as LoadIcon} from './IconList/LoadIcon';
+export {default as EditIcon} from './IconList/EditIcon';
+export {default as HomeIcon} from './IconList/HomeIcon';
+export {default as FilesIcon} from './IconList/FilesIcon';
+export {default as ProfileIcon} from './IconList/ProfileIcon';
+export {default as DocumentIcon} from './IconList/DocumentIcon';
+export {default as DeleteIcon} from './IconList/DeleteIcon';
+export {default as ContentCopyIcon} from './IconList/ContentCopyIcon';
+export {default as PaymentMethodIcon} from './IconList/PaymentMethodIcon';
+export {default as SocialIcon} from './IconList/SocialIcon';
+export {default as CloseIcon} from './IconList/CloseIcon';
+export {default as SearchIcon} from './IconList/SearchIcon';
